Fix merge conflict issues
