+++ src/main/webapp/app/core/user/settings/passkey-settings/passkey-settings.component.ts
@@ -17,20 +17,6 @@ import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 import { CustomMaxLengthDirective } from 'app/shared/validators/custom-max-length-validator/custom-max-length-validator.directive';
 import { addNewPasskey } from 'app/core/user/settings/passkey-settings/util/credential.util';
-import { createCredentialOptions } from 'app/core/user/settings/passkey-settings/util/credential-option.util';
-import { getCredentialWithGracefullyHandlingAuthenticatorIssues } from 'app/core/user/settings/passkey-settings/util/credential.util';
-import { InvalidCredentialError } from 'app/core/user/settings/passkey-settings/entities/invalid-credential-error';
-import { getOS } from 'app/shared/util/os-detector.util';
-
-const InvalidStateError = {
-    name: 'InvalidStateError',
-    authenticatorCredentialAlreadyRegisteredWithRelyingPartyCode: 11,
-};
-
-const UserAbortedPasskeyCreationError = {
-    code: 0,
-    name: 'NotAllowedError',
-};
 
 export interface DisplayedPasskey extends PasskeyDTO {
     isEditingLabel?: boolean;
@@ -93,51 +79,6 @@ export class PasskeySettingsComponent implements OnDestroy {
 
     async updateRegisteredPasskeys(): Promise<void> {
         this.registeredPasskeys.set(await this.passkeySettingsApiService.getRegisteredPasskeys());
-    }
-
-    async addNewPasskey() {
-        try {
-            const user = this.currentUser();
-            if (!user) {
-                // noinspection ExceptionCaughtLocallyJS - intended to be caught locally
-                throw new Error('User or Username is not defined');
-            }
-            const registrationOptions = await this.webauthnApiService.getRegistrationOptions();
-            const credentialOptions = createCredentialOptions(registrationOptions, user);
-
-            const authenticatorCredential = await navigator.credentials.create({
-                publicKey: credentialOptions,
-            });
-            const credential = getCredentialWithGracefullyHandlingAuthenticatorIssues(authenticatorCredential);
-            if (!credential) {
-                // noinspection ExceptionCaughtLocallyJS - intended to be caught locally
-                throw new InvalidCredentialError();
-            }
-
-            await this.webauthnApiService.registerPasskey({
-                publicKey: {
-                    credential: credential,
-                    label: `${user.email} - ${getOS()}`,
-                },
-            });
-        } catch (error) {
-            const userPressedCancelInPasskeyCreationDialog = error.name == UserAbortedPasskeyCreationError.name && error.code == UserAbortedPasskeyCreationError.code;
-            if (userPressedCancelInPasskeyCreationDialog) {
-                return;
-            }
-
-            if (error instanceof InvalidCredentialError) {
-                this.alertService.addErrorAlert('artemisApp.userSettings.passkeySettingsPage.error.invalidCredential');
-            } else if (error.name == InvalidStateError.name && error.code == InvalidStateError.authenticatorCredentialAlreadyRegisteredWithRelyingPartyCode) {
-                this.alertService.addErrorAlert('artemisApp.userSettings.passkeySettingsPage.error.passkeyAlreadyRegistered');
-            } else {
-                this.alertService.addErrorAlert('artemisApp.userSettings.passkeySettingsPage.error.registration');
-            }
-
-            throw error;
-        }
-
-        await this.updateRegisteredPasskeys();
     }
 
     private loadCurrentUser() {
