+++ src/main/webapp/app/core/user/settings/passkey-settings/util/credential.util.ts
@@ -58,12 +58,17 @@ export async function addNewPasskey(user: User | undefined, webauthnApiService: 
             // noinspection ExceptionCaughtLocallyJS - intended to be caught locally
             throw new Error('User or Username is not defined');
         }
-        const options = await webauthnApiService.getRegistrationOptions();
+        const registrationOptions = await webauthnApiService.getRegistrationOptions();
+        const credentialOptions = createCredentialOptions(registrationOptions, user);
 
-        const credentialOptions = createCredentialOptions(options, user);
-        const credential = await navigator.credentials.create({
+        const authenticatorCredential = await navigator.credentials.create({
             publicKey: credentialOptions,
         });
+        const credential = getCredentialWithGracefullyHandlingAuthenticatorIssues(authenticatorCredential);
+        if (!credential) {
+            // noinspection ExceptionCaughtLocallyJS - intended to be caught locally
+            throw new InvalidCredentialError();
+        }
 
         await webauthnApiService.registerPasskey({
             publicKey: {
@@ -72,15 +77,19 @@ export async function addNewPasskey(user: User | undefined, webauthnApiService: 
             },
         });
     } catch (error) {
-        if (error.name == UserAbortedPasskeyCreationError.name && error.code == UserAbortedPasskeyCreationError.code) {
-            return; // the user pressed cancel in the passkey creation dialog
+        const userPressedCancelInPasskeyCreationDialog = error.name == UserAbortedPasskeyCreationError.name && error.code == UserAbortedPasskeyCreationError.code;
+        if (userPressedCancelInPasskeyCreationDialog) {
+            return;
         }
 
-        if (error.name == InvalidStateError.name && error.code == InvalidStateError.authenticatorCredentialAlreadyRegisteredWithRelyingPartyCode) {
+        if (error instanceof InvalidCredentialError) {
+            alertService.addErrorAlert('artemisApp.userSettings.passkeySettingsPage.error.invalidCredential');
+        } else if (error.name == InvalidStateError.name && error.code == InvalidStateError.authenticatorCredentialAlreadyRegisteredWithRelyingPartyCode) {
             alertService.addErrorAlert('artemisApp.userSettings.passkeySettingsPage.error.passkeyAlreadyRegistered');
         } else {
             alertService.addErrorAlert('artemisApp.userSettings.passkeySettingsPage.error.registration');
         }
-        return;
+
+        throw error;
     }
 }
